project/gcf-facial-recognition/requirements.txt
functions-framework==3.*
google-cloud-vision
flask